Clean and complete tests (#184)
* clean and complete tests

* remove whitespace

* add kl_sym in symmetry test

* remove last warning on lagged coherence

* simplify test for coherence

* improve kl div and complete doc

* doc

* use scipy solver instead

* replace np.dot by @

* remove numpy import from utils.geodesic

* simplify notations

+++ pyriemann/stats.py
@@ -428,7 +428,7 @@ class PermutationDistance(BasePermutation):
         pattern = np.zeros((n_samples, n_samples))
         for classe in classes:
             ix = (y == classe)
-            pattern += (np.outer(ix, ix) / np.float(ix.sum()))
+            pattern += (np.outer(ix, ix) / ix.sum())
 
         within_ss = (X * pattern).sum() / 2
